Int to Float Converted to float so that more things use RadiationEntry

+++ src/main/java/net/timeworndevs/curieapi/util/ArmorInsulator.java
@@ -39,7 +39,7 @@ public record ArmorInsulator (ArrayList<Item> armorItems, ArrayList<Float> multi
 
     }
 
-    public int getRadiation(RadiationType type) {
+    public float getRadiation(RadiationType type) {
         return this.radiations.get(type);
     }
 }
